# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WTS Machines API",
    "version": "1.0.0",
    "description": "Backend für Maschinen-Monitoring mit Azure SQL und Socket.IO."
  },
  "servers": [
    { "url": "https://{host}", "variables": { "host": { "default": "example.com" } } }
  ],
  "tags": [
    { "name": "Health" },
    { "name": "Machines" }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Liveness",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ok" },
                "examples": { "ok": { "value": { "ok": true } } }
              }
            }
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "tags": ["Health"],
        "summary": "Readiness (DB)",
        "responses": {
          "200": {
            "description": "DB erreichbar",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ok" },
                "examples": { "ok": { "value": { "ok": true } } }
              }
            }
          },
          "503": {
            "description": "DB nicht erreichbar",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ok" },
                "examples": { "down": { "value": { "ok": false } } }
              }
            }
          }
        }
      }
    },
    "/api/machines/basic": {
      "get": {
        "tags": ["Machines"],
        "summary": "Liste kompakter Maschinendaten",
        "responses": {
          "200": {
            "description": "Liste",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "machines": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MachineBasic" }
                    }
                  },
                  "required": ["machines"]
                },
                "examples": {
                  "example": {
                    "value": {
                      "machines": [
                        {
                          "name": "Maschine1",
                          "identifikation": "UG39BE-47H8UJ",
                          "letzteWartung": "01.08.2017",
                          "durchgängigeLaufzeit": 2059
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/machines/{name}": {
      "get": {
        "tags": ["Machines"],
        "summary": "Maschinen-Details",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MachineDetail" },
                "examples": {
                  "example": {
                    "value": {
                      "identifikation": "UG39BE-47H8UJ",
                      "letzteWartung": "01.08.2017",
                      "durchgängigeLaufzeit": 2059,
                      "temperatur": 28.12,
                      "aktuelleLeistung": 56,
                      "betriebsminutenGesamt": 20059,
                      "geschwindigkeit": 2.15
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": { "notFound": { "value": { "error": "not found" } } }
              }
            }
          }
        }
      }
    },
    "/api/machines/{name}/telemetry": {
      "get": {
        "tags": ["Machines"],
        "summary": "Telemetry-Historie",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "since", "in": "query", "required": false, "schema": { "type": "string", "format": "date-time" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 } }
        ],
        "responses": {
          "200": {
            "description": "Historie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/TelemetryItem" } }
                  },
                  "required": ["items"]
                },
                "examples": {
                  "example": {
                    "value": {
                      "items": [
                        {
                          "ts": "2025-08-29T12:00:00Z",
                          "temperatur": 28.4,
                          "aktuelleLeistung": 55,
                          "betriebsminutenGesamt": 20060,
                          "geschwindigkeit": 2.10
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nicht gefunden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": { "notFound": { "value": { "error": "not found" } } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Machines"],
        "summary": "Telemetry einfügen",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TelemetryPost" },
              "examples": {
                "partial": {
                  "value": { "temperatur": 42.1, "aktuelleLeistung": 63 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Erstellt",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ok" },
                "examples": { "ok": { "value": { "ok": true } } }
              }
            }
          },
          "400": {
            "description": "Ungültiger Body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": { "bad": { "value": { "error": "invalid" } } }
              }
            }
          },
          "404": {
            "description": "Maschine nicht gefunden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": { "notFound": { "value": { "error": "not_found" } } }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ok": {
        "type": "object",
        "properties": { "ok": { "type": "boolean" } },
        "required": ["ok"]
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
      "MachineBasic": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "identifikation": { "type": "string" },
          "letzteWartung": { "type": "string" },
          "durchgängigeLaufzeit": { "type": "number" }
        },
        "required": ["name", "identifikation", "letzteWartung", "durchgängigeLaufzeit"]
      },
      "MachineDetail": {
        "type": "object",
        "properties": {
          "identifikation": { "type": "string" },
          "letzteWartung": { "type": "string" },
          "durchgängigeLaufzeit": { "type": "number" },
          "temperatur": { "type": "number" },
          "aktuelleLeistung": { "type": "number", "minimum": 0, "maximum": 100 },
          "betriebsminutenGesamt": { "type": "number", "minimum": 0 },
          "geschwindigkeit": { "type": "number", "minimum": 0 }
        },
        "required": ["identifikation", "letzteWartung", "durchgängigeLaufzeit", "temperatur", "aktuelleLeistung", "betriebsminutenGesamt", "geschwindigkeit"]
      },
      "TelemetryItem": {
        "type": "object",
        "properties": {
          "ts": { "type": "string", "format": "date-time" },
          "temperatur": { "type": "number" },
          "aktuelleLeistung": { "type": "number", "minimum": 0, "maximum": 100 },
          "betriebsminutenGesamt": { "type": "number", "minimum": 0 },
          "geschwindigkeit": { "type": "number", "minimum": 0 }
        },
        "required": ["ts"]
      },
      "TelemetryPost": {
        "type": "object",
        "properties": {
          "temperatur": { "type": "number" },
          "aktuelleLeistung": { "type": "number", "minimum": 0, "maximum": 100 },
          "betriebsminutenGesamt": { "type": "number", "minimum": 0 },
          "geschwindigkeit": { "type": "number", "minimum": 0 }
        },
        "additionalProperties": false,
        "minProperties": 1
      }
    }
  }
}